refactor(gear): share gem/enchant option helpers
