+++ frontend/src/app/components/gear/GemSelector.tsx
@@ -9,11 +9,7 @@ import { useWowheadTooltips } from '../../lib/useWowheadTooltips';
 import { useLanguage } from '../../lib/i18n';
 import Switch from '../ui/Switch';
 import CollapsibleSection from '../ui/CollapsibleSection';
-import { statLabel, type ItemOption } from './itemOptions';
-
-interface GemOption extends ItemOption {
-  algariColor?: string;
-}
+import { statLabel, filterDiamonds, groupGemsByColor, type GemOption } from './itemOptions';
 
 interface GemSelectorProps {
   equippedSlots: Record<string, ResolvedItem>;
@@ -84,28 +80,10 @@ export default function GemSelector({
   }, [hasSocketedSlots]);
 
   // Diamonds = quality 4, crafted rank 2 (separate from regular gems)
-  const diamonds = useMemo(
-    () => gemOptions.filter((g) => g.craftingQuality === 2 && (g.quality ?? 0) === 4),
-    [gemOptions]
-  );
+  const diamonds = useMemo(() => filterDiamonds(gemOptions), [gemOptions]);
 
   // Regular gems grouped by color: rank 2 crafted, quality 3 (Flawless rare)
-  const gemGroups = useMemo(() => {
-    const filtered = gemOptions.filter((g) => g.craftingQuality === 2 && (g.quality ?? 0) === 3);
-    const groups: Record<string, GemOption[]> = {};
-    for (const g of filtered) {
-      const color = g.algariColor || 'other';
-      if (!groups[color]) groups[color] = [];
-      groups[color].push(g);
-    }
-    for (const arr of Object.values(groups)) {
-      arr.sort((a, b) => (a.itemName || a.displayName).localeCompare(b.itemName || b.displayName));
-    }
-    const colorOrder = ['amethyst', 'garnet', 'lapis', 'peridot', 'other'];
-    return colorOrder
-      .filter((c) => groups[c]?.length > 0)
-      .map((c) => ({ color: c, gems: groups[c] }));
-  }, [gemOptions]);
+  const gemGroups = useMemo(() => groupGemsByColor(gemOptions), [gemOptions]);
 
   const allGemIds = useMemo(
     () => gemGroups.flatMap((g) => g.gems.map((gem) => gem.itemId!).filter(Boolean)),
